Extract slack service to common directory
