+++ common/services/slack/security.js
