+++ common/services/slack/shortcuts.js
