+++ lib/controllers/slack.js
@@ -1,9 +1,9 @@
-const commands = require('../services/slack/commands');
-const shortcuts = require('../services/slack/shortcuts');
-const viewSubmissions = require('../services/slack/view-submissions');
+const commands = require('../../common/services/slack/commands');
+const shortcuts = require('../../common/services/slack/shortcuts');
+const viewSubmissions = require('../../common/services/slack/view-submissions');
 const github = require('../services/github');
 const googleSheet = require('../services/google-sheet');
-const postSlackMessage = require('../services/slack/surfaces/messages/post-message');
+const postSlackMessage = require('../../common/services/slack/surfaces/messages/post-message');
 
 function _getDeployStartedMessage(release, appName) {
   return `Commande de déploiement de la release "${release}" pour ${appName} en production bien reçue.`;
