+++ lib/routes/slack.js
@@ -1,4 +1,4 @@
-const { verifySignatureAndParseBody } = require('../services/slack/security');
+const { verifySignatureAndParseBody } = require('../../common/services/slack/security');
 const slackbotController = require('../controllers/slack');
 
 const slackConfig = {
